Move providers to hub, add api key to agixt
